requirements.txt
# Core dependencies
python-dateutil>=2.8.2
pandas>=1.3.0

# Google Cloud dependencies
google-cloud-documentai>=2.13.0
google-api-core>=2.8.0
vertexai>=0.1.0

# Snowflake
snowflake-connector-python>=2.7.0

# Email processing
# Note: imaplib is part of the Python standard library
# For Outlook support (optional)
exchangelib>=4.0.0
requests>=2.25.0